version: update to 26.1
