+++ .gitignore
@@ -33,5 +33,3 @@ bin/
 run/
 
 #own
-
-assets/
