+++ gradle.properties
@@ -3,17 +3,17 @@ org.gradle.jvmargs=-Xmx1G
 
 # Fabric Properties
 # check these on https://fabricmc.net/versions.html
-minecraft_version=1.21.11
-yarn_mappings=1.21.11+build.2
-loader_version=0.18.2
+minecraft_version=26.1-pre-2
+loader_version=0.18.4
+loom_version=1.15-SNAPSHOT
 
-vanillaconfig_version = 1.2.15+1.21.9
-mod_menu_version = 14.0.0-rc.2
+vanillaconfig_version = 1.2.16+26.1-beta.4
+mod_menu_version = 18.0.0-alpha.5
 
 # Mod Properties
-mod_version = 1.0.11+1.21.11
+mod_version = 1.0.12+26.1-beta.1
 maven_group = net.treset
 archives_base_name = audio_hotkeys
 
 # Dependencies
-fabric_version=0.139.4+1.21.11
+fabric_api_version=0.143.14+26.1
