+++ src/client/java/net/treset/audio_hotkeys/mixin/GuiMixin.java
@@ -1,0 +1,19 @@
+package net.treset.audio_hotkeys.mixin;
+
+import net.minecraft.client.DeltaTracker;
+import net.minecraft.client.gui.Gui;
+import net.minecraft.client.gui.GuiGraphicsExtractor;
+import net.treset.audio_hotkeys.gui.OverlayManager;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Inject;
+import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
+
+@Mixin(Gui.class)
+public class GuiMixin {
+
+    @Inject(method = "extractRenderState(Lnet/minecraft/client/gui/GuiGraphicsExtractor;Lnet/minecraft/client/DeltaTracker;)V", at = @At("TAIL"), cancellable = true)
+    public void onRender(GuiGraphicsExtractor ctx, DeltaTracker tickDelta, CallbackInfo info) {
+        OverlayManager.renderOverlay(ctx);
+    }
+}
