+++ src/client/resources/assets/audio_hotkeys/icon.png
