+++ src/client/resources/assets/audio_hotkeys/lang/en_us.json
@@ -1,0 +1,119 @@
+{
+  "config.audiohotkeys.main.page": "Audio Hotkeys Settings",
+  "config.audiohotkeys.master.page": "Master Hotkeys",
+  "config.audiohotkeys.music.page": "Music Hotkeys",
+  "config.audiohotkeys.ingame.page": "InGame Hotkeys",
+  "config.audiohotkeys.creatures.page": "Creature Hotkeys",
+  "config.audiohotkeys.other.page": "Other Hotkeys",
+
+  "config.audiohotkeys.open_config": "Open Settings",
+  "config.audiohotkeys.open_config.comment": "Hotkey that opens this menu.",
+
+  "config.audiohotkeys.master.mute": "Mute Master",
+  "config.audiohotkeys.master.mute.comment": "Mute all Game Audio.",
+  "config.audiohotkeys.master.up": "Master Up",
+  "config.audiohotkeys.master.up.comment": "Increase the general Volume.",
+  "config.audiohotkeys.master.down": "Master Down",
+  "config.audiohotkeys.master.down.comment": "Decrease the general Volume.",
+  "config.audiohotkeys.master.step": "Master Step",
+  "config.audiohotkeys.master.step.comment": "Amount to Increase / Decrease the general Volume by.",
+  "config.audiohotkeys.subtitles.mute": "Toggle Subtitles",
+  "config.audiohotkeys.subtitles.mute.comment": "Toggle the InGame Subtitles.",
+
+  "config.audiohotkeys.music.mute": "Mute Music",
+  "config.audiohotkeys.music.mute.comment": "Mute the background Music.",
+  "config.audiohotkeys.music.up": "Music Up",
+  "config.audiohotkeys.music.up.comment": "Increase the background Music Volume.",
+  "config.audiohotkeys.music.down": "Music Down",
+  "config.audiohotkeys.music.down.comment": "Decrease the background Music Volume.",
+  "config.audiohotkeys.music.step": "Music Step",
+  "config.audiohotkeys.music.step.comment": "Amount to Increase / Decrease the background Music Volume by.",
+  "config.audiohotkeys.jukebox.mute": "Mute Jukebox",
+  "config.audiohotkeys.jukebox.mute.comment": "Mute Jukeboxes and Note Blocks.",
+  "config.audiohotkeys.jukebox.up": "Jukebox Up",
+  "config.audiohotkeys.jukebox.up.comment": "Increase the Jukebox and Note Block Volume.",
+  "config.audiohotkeys.jukebox.down": "Jukebox Down",
+  "config.audiohotkeys.jukebox.down.comment": "Decrease the Jukebox and Note Block Volume.",
+  "config.audiohotkeys.jukebox.step": "Jukebox Step",
+  "config.audiohotkeys.jukebox.step.comment": "Amount to Increase / Decrease the Jukebox and Note Block Volume by.",
+
+  "config.audiohotkeys.environment.mute": "Mute Environment",
+  "config.audiohotkeys.environment.mute.comment": "Mute Environmental Audio",
+  "config.audiohotkeys.environment.up": "Environment Up",
+  "config.audiohotkeys.environment.up.comment": "Increase the Environmental Audio Volume.",
+  "config.audiohotkeys.environment.down": "Environment Down",
+  "config.audiohotkeys.environment.down.comment": "Decrease the Environmental Audio Volume.",
+  "config.audiohotkeys.environment.step": "Environment Step",
+  "config.audiohotkeys.environment.step.comment": "Amount to Increase / Decrease the Environmental Audio Volume by.",
+  "config.audiohotkeys.weather.mute": "Mute Weather",
+  "config.audiohotkeys.weather.mute.comment": "Mute Weather Sounds",
+  "config.audiohotkeys.weather.up": "Weather Up",
+  "config.audiohotkeys.weather.up.comment": "Increase the Weather Volume.",
+  "config.audiohotkeys.weather.down": "Weather Down",
+  "config.audiohotkeys.weather.down.comment": "Decrease the Weather Volume.",
+  "config.audiohotkeys.weather.step": "Weather Step",
+  "config.audiohotkeys.weather.step.comment": "Amount to Increase / Decrease the Weather Volume by.",
+  "config.audiohotkeys.blocks.mute": "Mute Blocks",
+  "config.audiohotkeys.blocks.mute.comment": "Mute Block Sounds",
+  "config.audiohotkeys.blocks.up": "Blocks Up",
+  "config.audiohotkeys.blocks.up.comment": "Increase the Block Volume.",
+  "config.audiohotkeys.blocks.down": "Blocks Down",
+  "config.audiohotkeys.blocks.down.comment": "Decrease the Block Volume.",
+  "config.audiohotkeys.blocks.step": "Blocks Step",
+  "config.audiohotkeys.blocks.step.comment": "Amount to Increase / Decrease the Block Volume by.",
+
+  "config.audiohotkeys.hostile.mute": "Mute Hostile Mobs",
+  "config.audiohotkeys.hostile.mute.comment": "Mute Hostile Mob Sounds",
+  "config.audiohotkeys.hostile.up": "Hostile Mobs Up",
+  "config.audiohotkeys.hostile.up.comment": "Increase the Hostile Mob Volume.",
+  "config.audiohotkeys.hostile.down": "Hostile Mobs Down",
+  "config.audiohotkeys.hostile.down.comment": "Decrease the Hostile Mob Volume.",
+  "config.audiohotkeys.hostile.step": "Hostile Mobs Step",
+  "config.audiohotkeys.hostile.step.comment": "Amount to Increase / Decrease the Hostile Mob Volume by.",
+  "config.audiohotkeys.friendly.mute": "Mute Friendly Mobs",
+  "config.audiohotkeys.friendly.mute.comment": "Mute Friendly Mob Sounds",
+  "config.audiohotkeys.friendly.up": "Friendly Mobs Up",
+  "config.audiohotkeys.friendly.up.comment": "Increase the Friendly Mob Volume.",
+  "config.audiohotkeys.friendly.down": "Friendly Mobs Down",
+  "config.audiohotkeys.friendly.down.comment": "Decrease the Friendly Mob Volume.",
+  "config.audiohotkeys.friendly.step": "Friendly Mobs Step",
+  "config.audiohotkeys.friendly.step.comment": "Amount to Increase / Decrease the Friendly Mob Volume by.",
+  "config.audiohotkeys.player.mute": "Mute Players",
+  "config.audiohotkeys.player.mute.comment": "Mute Player Sounds.",
+  "config.audiohotkeys.player.up": "Players Up",
+  "config.audiohotkeys.player.up.comment": "Increase the Player Volume.",
+  "config.audiohotkeys.player.down": "Players Down",
+  "config.audiohotkeys.player.down.comment": "Decrease the Player Volume.",
+  "config.audiohotkeys.player.step": "Players Step",
+  "config.audiohotkeys.player.step.comment": "Amount to Increase / Decrease the Player Volume by.",
+
+  "config.audiohotkeys.voice.mute": "Mute Voice",
+  "config.audiohotkeys.voice.mute.comment": "Mute Voice Audio (Narrator).",
+  "config.audiohotkeys.voice.up": "Voice Up",
+  "config.audiohotkeys.voice.up.comment": "Increase the Voice Audio Volume.",
+  "config.audiohotkeys.voice.down": "Voice Down",
+  "config.audiohotkeys.voice.down.comment": "Decrease the Voice Audio Volume.",
+  "config.audiohotkeys.voice.step": "Voice Step",
+  "config.audiohotkeys.voice.step.comment": "Amount to Increase / Decrease the Voice Audio Volume by.",
+
+  "state.audiohotkeys.master": "Master",
+  "state.audiohotkeys.subtitles": "Subtitles",
+  "state.audiohotkeys.music": "Music",
+  "state.audiohotkeys.jukebox": "Jukebox",
+  "state.audiohotkeys.environment": "Environment",
+  "state.audiohotkeys.weather": "Weather",
+  "state.audiohotkeys.blocks": "Blocks",
+  "state.audiohotkeys.hostile": "Hostile Mobs",
+  "state.audiohotkeys.friendly": "Friendly Mobs",
+  "state.audiohotkeys.player": "Player",
+  "state.audiohotkeys.voice": "Voice",
+
+  "string.audiohotkeys.mute": "Muted %s!",
+  "string.audiohotkeys.unmute": "Unmuted %s. Volume is now %s%%!",
+  "string.audiohotkeys.unmute_failed": "Unmuted %s but didn't change the manually set value of %s%%!",
+  "string.audiohotkeys.increase": "Increased %s volume by %s%% to %s%%!",
+  "string.audiohotkeys.decrease": "Decreased %s volume by %s%% to %s%%!",
+  "string.audiohotkeys.activate": "Activated %s!",
+  "string.audiohotkeys.deactivate": "Deactivated %s!"
+
+}
