--- src/main/java/net/treset/audio_hotkeys/mixin/InGameHudMixin.java
@@ -1,19 +1,0 @@
-package net.treset.audio_hotkeys.mixin;
-
-import net.minecraft.client.gui.DrawContext;
-import net.minecraft.client.gui.hud.InGameHud;
-import net.minecraft.client.render.RenderTickCounter;
-import net.treset.audio_hotkeys.gui.OverlayManager;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
-
-@Mixin(InGameHud.class)
-public class InGameHudMixin {
-
-    @Inject(method = "render(Lnet/minecraft/client/gui/DrawContext;Lnet/minecraft/client/render/RenderTickCounter;)V", at = @At("TAIL"), cancellable = true)
-    public void onRender(DrawContext ctx, RenderTickCounter tickDelta, CallbackInfo info) {
-        OverlayManager.renderOverlay(ctx);
-    }
-}
