--- src/main/java/net/treset/audio_hotkeys/tools/MathTools.java
@@ -1,7 +1,0 @@
-package net.treset.audio_hotkeys.tools;
-
-public class MathTools {
-    public static float clamp(float val, float min, float max) {
-        return Math.max(min, Math.min(max, val));
-    }
-}
