--- src/main/resources/audio_hotkeys.mixins.json
@@ -1,14 +1,0 @@
-{
-  "required": true,
-  "minVersion": "0.8",
-  "package": "net.treset.audio_hotkeys.mixin",
-  "compatibilityLevel": "JAVA_16",
-  "mixins": [
-  ],
-  "client": [
-    "InGameHudMixin"
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
