+++ src/main/resources/fabric.mod.json
@@ -30,14 +30,17 @@
     ]
   },
   "mixins": [
-    "audio_hotkeys.mixins.json"
+      {
+          "config": "audio_hotkeys.client.mixins.json",
+          "environment": "client"
+      }
   ],
 
   "depends": {
     "fabricloader": ">=0.14.20",
-    "vanillaconfig": ">=1.2.12",
-    "minecraft": ["~1.21.6"],
-    "java": ">=21"
+    "vanillaconfig": ">=1.2.16-",
+    "minecraft": ["~26.1-"],
+    "java": ">=25"
   },
   "suggests": {
     "modmenu": "*"
